requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
airtable-python-wrapper==0.15.1
antlr4-python3-runtime==4.9.1
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
aws-shell==0.2.2
awscli==1.19.16
black==20.8b1
blessed==1.17.6
blessings==1.7
boto3==1.17.16
botocore==1.20.16
bpython==0.21
CacheControl==0.12.6
cachetools==4.2.1
cachy==0.3.0
Cerberus==1.3.2
certifi==2020.12.5
cffi==1.14.5
chalice==1.22.1
chardet==4.0.0
cleo==0.7.6
click==7.1.2
clikit==0.4.3
colorama==0.4.3
configobj==5.0.6
crashtest==0.3.1
cryptography==3.4.4
cs50==6.0.2
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
dephell==0.8.3
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.4
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
distlib==0.3.1
docopt==0.6.2
docutils==0.15.2
enum-compat==0.0.3
filelock==3.0.12
Flask==1.1.2
funcsigs==1.0.2
gast==0.3.3
glfw==2.0.0
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.35.0
h5py==2.10.0
html5lib==1.1
idna==2.10
inquirer==2.7.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==20.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
m2r==0.2.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
parso==0.7.1
pastel==0.2.1
pathspec==0.8.1
pexpect==4.8.0
Pillow==8.1.0
pipreqs-amasad==0.4.10
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.0.5
poetry-core==1.0.2
prompt-toolkit==1.0.18
protobuf==3.14.0
ptpython==3.0.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pygame==2.0.1
Pygments==2.7.4
pylev==1.3.0
pylint==2.6.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyquil==1.9.0
pyrsistent==0.14.11
python-dateutil==2.8.1
python-editor==1.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pyxdg==0.27
pyxel==1.0.1
PyYAML==5.4.1
readchar==2.0.1
referenceqvm==0.3
regex==2020.11.13
replit==2.0.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rope==0.18.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
scipy==1.4.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.15.0
sounddevice==0.4.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.5.11
tqdm==4.56.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
yarg==0.1.9
yarl==1.6.3
yaspin==1.3.0